Dashboard: use new style for registration
The API currently used by the demo application is deprecated and the
server warns about this upon startup:

  DEPRECATION WARNING: ActiveAdmin::Dashboard is deprecated and will be
removed in the next version

Fix this by using the new API, and amend the documentation accordingly.

Note that the new page that is generated has a different layout. I
decided to move the graphical chart to a new row, as the automatic width
distribution didn't make the surrounding div adopt to the content width
of 500px anymore, which resulted in a cropped image display. I don't know
if there's a way to override that behaviour.