disable comments
We receive ~200 comments per minute. And Heroku tels us about turning of our database.

```
[warning] Database disruption imminent, row limit exceeded for dev database on Heroku app demo-activeadmin 5 days ago
The database HEROKU_POSTGRESQL_GOLD_URL on Heroku app demo-activeadmin has exceeded its allocated storage capacity. Immediate action is required.

The database contains 52,421 rows, exceeding the Dev plan limit of 10,000. INSERT privileges to the database will be automatically revoked in 7 days. This will cause service failures in most applications dependent on this database.

To avoid a disruption to your service, migrate the database to a Hobby Basic ($9/month) or higher database plan:

https://devcenter.heroku.com/articles/upgrade-heroku-postgres-with-pgbackups

If you are unable to upgrade the database, you should reduce the number of records stored in it.
```

We need less then 2 hours to jump over the 10k limit.