Fully reload DB on deployments
I was expecting heroku to run the postdeploy script on regular builds
too, but it seems like it's only run for review apps, and we need to
configure it in the `Procfile` to be run on regular builds.

https://devcenter.heroku.com/articles/release-phase

Since our DB is ephemeral, we can drop our postdeploy script and reseed
everything when releasing.