Remove mechanize gem and products rake task (#775)
* Bump TargetRubyVersion from 3.1 to 3.2

* Remove products rake task, doesn't work now

These URLs fail with a 503 despite being accessible. We have the sample data we need.

* Remove mechanize gem

* Move erb_lint to rubocop section

We want this grouped with rubocop since it works with it.

* bundle update --bundler

* Remove unused gems

We'll add simplecov back when we decide to integrate Codecov in this repo. I don't think we'd need webmock as plain request specs would suffice.

* Move faker to development/test group

This is only used in db/seeds so no need to always install it. We only need it in development or test environments.

* Trigger GHA lint required workflow

Path filtering doesn't work in public beta of required workflows