Temporary v4 conversion for testing in Render
We want to test a deployed app on v4 in a production environment. We'll use Render for that.

Note that the sessions approach here is quite bad and broken. You can bypass it entirely. We aren't using admin_user model but user model instead with no devise actions so there isn't any sort of protection or integration. We'll revisit this when we rebuild the demo app.