Disable the logged sqlite3 production warning
We want to use sqlite3 in production because its simplest to setup and use for the demo.

> W, [2024-01-06T15:05:37.098154 #61]  WARN -- : You are running SQLite in production, this is generally not recommended. You can disable this warning by setting "config.active_record.sqlite3_production_warning=false".

No need for Postgres. If anything in the future, we can change to use Postgres in production but in development continue with sqlite3. We normally wouldn't do that but since its a demo app its very convenient to not need to rely on Postgres which would normally require using Docker Compose for easier setup. We avoid all that by using sqlite3.