Block updating and deleting default admin user
The batch actions for admin user are already disabled. We need to ensure a default admin user remains for access.

Also configures the default email in a constant. Most tests assertions we'll leave hard coded as that's best but when creating the admin user (or logging manually for one test case) we'll use the configured email. The main reason for this is we also have to reference it in db/seeds and the admin resource as the critical key.