Block updates to default admin user record (#820)
* Block updating and deleting default admin user

The batch actions for admin user are already disabled. We need to ensure a default admin user remains for access.

Also configures the default email in a constant. Most tests assertions we'll leave hard coded as that's best but when creating the admin user (or logging manually for one test case) we'll use the configured email. The main reason for this is we also have to reference it in db/seeds and the admin resource as the critical key.

* Add coverage for default update/destroy actions

* Remove coverage settings to allow upload

With these settings, simplecov returns a non-zero exit status which we don't want because it fails the test run step in the CI job and never gets to uploading the code coverage report for us to see what's wrong in the PR. For what its worth, I haven't run into this before with rspec so not sure if something could be different here with simplecov and rails test.