Improve setup instructions
I found that it was missing that you need `yarn` installed, and `yarn
install` run in order to get started.

To make `yarn` available, I added a `packageManager` field with latest
yarn 1.x version to `package.json` and included `corepack enable` so
that this field is respected and nodejs automatically installs yarn when
first run.

I also added `yarn install` to setup instructions.